refactor(provider/amazon): Convert createClassicLoadBalancer to TS and extract common functions
