+++ app/scripts/modules/amazon/src/loadBalancer/loadBalancer.module.ts
@@ -1,6 +1,7 @@
 import { module } from 'angular';
 
 import { AWS_CREATE_APPLICATION_LOAD_BALANCER_CTRL } from './configure/application/createApplicationLoadBalancer.controller';
+import { AWS_CREATE_CLASSIC_LOAD_BALANCER_CTRL } from './configure/classic/createClassicLoadBalancer.controller';
 import { AWS_LOAD_BALANCER_CHOICE_MODAL } from './configure/choice/awsLoadBalancerChoice.modal';
 import { AWS_LOAD_BALANCER_DETAILS_CTRL } from './details/loadBalancerDetails.controller';
 import { AWS_LOAD_BALANCER_TRANFORMER } from './loadBalancer.transformer';
@@ -10,7 +11,7 @@ import { TARGET_GROUP_STATES } from './targetGroup.states';
 export const AWS_LOAD_BALANCER_MODULE = 'spinnaker.amazon.loadBalancer';
 
 module(AWS_LOAD_BALANCER_MODULE, [
-  require('./configure/classic/createClassicLoadBalancer.controller'),
+  AWS_CREATE_CLASSIC_LOAD_BALANCER_CTRL,
   AWS_CREATE_APPLICATION_LOAD_BALANCER_CTRL,
   AWS_LOAD_BALANCER_CHOICE_MODAL,
   AWS_LOAD_BALANCER_DETAILS_CTRL,
